bundler: copy the plugin handle into the hops; drop the dead drainDeferred reject path
Resolve, Load and DeferredBatchTask now carry BundleV2::plugins, copied
when they are built on the bundle thread, the way Resolve and Load already
copy everything else their plugin-thread body needs. The bodies read only
their own fields, so the raw-pointer accessor and the from_field_ptr!
walk-back from the previous commit go away, along with DeferredBatchTask's
never-set debug flag.

The deferred batch also read completion.result_is_err() to decide whether
to reject the .defer() promises. result stays Pending until after the pass
returns, and the batch only ever runs while the pass is waiting on the
deferred loads, so it was always false (bake passes have no completion at
all); it was also a read of a struct the bundle thread holds &mut to for
the whole pass. Remove it end to end: the vtable entry, the handle method,
and drainDeferred's rejected parameter.

The lint now bans any route from a hop body back to the pass.
