+++ src/jsc/bindings/JSBundlerPlugin.cpp
@@ -665,7 +665,7 @@ extern "C" void JSBundlerPlugin__setConfig(Bun::JSBundlerPlugin* plugin, void* c
     plugin->plugin.config = config;
 }
 
-extern "C" void JSBundlerPlugin__drainDeferred(Bun::JSBundlerPlugin* pluginObject, bool rejected)
+extern "C" void JSBundlerPlugin__drainDeferred(Bun::JSBundlerPlugin* pluginObject)
 {
     auto* globalObject = pluginObject->globalObject();
     MarkedArgumentBuffer arguments;
@@ -676,11 +676,7 @@ extern "C" void JSBundlerPlugin__drainDeferred(Bun::JSBundlerPlugin* pluginObjec
     auto scope = DECLARE_THROW_SCOPE(vm);
     for (auto promiseValue : arguments) {
         JSPromise* promise = uncheckedDowncast<JSPromise>(promiseValue);
-        if (rejected) {
-            promise->reject(vm, JSC::jsUndefined());
-        } else {
-            promise->resolve(globalObject, vm, JSC::jsUndefined());
-        }
+        promise->resolve(globalObject, vm, JSC::jsUndefined());
         RETURN_IF_EXCEPTION(scope, );
     }
     RETURN_IF_EXCEPTION(scope, );
